% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{add_snp_eff_annotations}
\alias{add_snp_eff_annotations}
\title{Add snpEff annotations to the variant table}
\usage{
add_snp_eff_annotations(var_table, snpeff_config)
}
\arguments{
\item{vcf_filename}{VCF file input}

\item{species}{Species corresponding to vcf_file}

\item{config}{Location of snpEff config file}
}
\value{
Path to the annotated snpEFF VCF file
}
\description{
Add snpEff annotations to the variant table
}
