% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{add_ts_tv_info}
\alias{add_ts_tv_info}
\title{Add transition/transversion annotation to variant table}
\usage{
add_ts_tv_info(var_table)
}
\arguments{
\item{var_table}{A variant table}
}
\value{
A variant table with ts_tv column appended
}
\description{
Add transition/transversion annotation to variant table
}
