% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{annotate_variant_table}
\alias{annotate_variant_table}
\title{Annotate a variant table with snpEff and ts/tv info}
\usage{
annotate_variant_table(var_table, snpeff_config)
}
\arguments{
\item{var_table}{A variant table}

\item{snpeff_config}{Path to the snpeff config file}
}
\description{
Annotate a variant table with snpEff and ts/tv info
}
