% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Wrapper script which performs the whole annotation pipeline}
\usage{
annotate_variants(var_table)
}
\arguments{
\item{var_table}{A variant table}
}
\value{
An annotated variant table
}
\description{
Wrapper script which performs the whole annotation pipeline
}
