% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_calling.R
\name{binom_tests}
\alias{binom_tests}
\title{Megadaph mitochondrial variant calling R library}
\usage{
binom_tests(mut_cov_count, seq_err)
}
\description{
Megadaph mitochondrial variant calling R library
}
