% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_mut_rate.R
\name{boot_diff}
\alias{boot_diff}
\title{Bootstrap difference in means between two groups}
\usage{
boot_diff(var_table, line_info, by, group1, group2)
}
\arguments{
\item{var_table}{A variant table}

\item{line_info}{A line info table}

\item{by}{Column name in var_table in which groups can be found}

\item{group1, group2}{Character; Group IDs in var_table}
}
\description{
Bootstrap difference in means between two groups
}
