% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{boot_mutation_rate}
\alias{boot_mutation_rate}
\title{Bootstrap the mutation rate statistic across samples}
\usage{
boot_mutation_rate(af, gen, nuc, reps = 1e+05)
}
\arguments{
\item{af}{list of numeric; mutant allele frequencies split by sample}

\item{gen}{list of numeric; mean generation number split by sample}

\item{nuc}{list of integer; bases surveyed split by sample}

\item{reps}{integer; Number of bootstrap replicates}
}
\value{
A bootstrapped sample of mutation rates
}
\description{
Bootstrap the mutation rate statistic across samples
}
