% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{boot_mutation_rate_diff_within}
\alias{boot_mutation_rate_diff_within}
\title{Bootstrap test for the equality of means}
\usage{
boot_mutation_rate_diff_within(var_table, line_info, by, within = NULL)
}
\arguments{
\item{var_table}{A variant table}

\item{line_info}{A line info table}

\item{by}{Group in var_table to calculate the difference between}

\item{within}{Group var_table to calculate the difference within}
}
\value{
Numeric; A bootstrap sample of differences in mean
}
\description{
https://stats.stackexchange.com/questions/136661/using-bootstrap-under-h0-to-perform-a-test-for-the-difference-of-two-means-repl
Efron's and Tibshirani's - intro to the bootstrap, page 223
}
