% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{bootstrap_mutation_rate_by}
\alias{bootstrap_mutation_rate_by}
\title{Bootstrap the mutation rate for a specific group}
\usage{
bootstrap_mutation_rate_by(group, by, variant_table, line_info, combined_mu,
  reps)
}
\description{
Bootstrap the mutation rate for a specific group
}
