% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{bootstrap_mutation_rate_quantiles}
\alias{bootstrap_mutation_rate_quantiles}
\title{Bootstrap the mutation rate across samples and return the quantiles}
\usage{
bootstrap_mutation_rate_quantiles(af, gen, nuc, reps = 1e+05)
}
\arguments{
\item{af}{list of numeric; mutant allele frequencies split by sample}

\item{gen}{list of numeric; mean generation number split by sample}

\item{nuc}{list of integer; bases surveyed split by sample}
}
\value{
Numeric vector; First item is the mutation rate. The
        other five are the 2.5%, 25%, 50%, 75%, and 97.5% quantiles.
}
\description{
Bootstrap the mutation rate across samples and return the quantiles
}
