% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{bootstrap_mutation_rate_test}
\alias{bootstrap_mutation_rate_test}
\title{Bootstrap difference in mutation rate between two groups}
\usage{
bootstrap_mutation_rate_test(variant_table, line_info, by, group1, group2,
  reps = 1e+05)
}
\arguments{
\item{variant_table}{A variant table}

\item{line_info}{A line info table}

\item{by}{Column name in variant_table in which groups can be found}

\item{group1, group2}{Character; Group IDs in variant_table}

\item{reps}{integer; Number of bootstrap replicates}
}
\value{
A p-value
}
\description{
#' https://stats.stackexchange.com/questions/136661/using-bootstrap-under-h0\
-to-perform-a-test-for-the-difference-of-two-means-repl
Efron's and Tibshirani's - intro to the bootstrap, page 223
}
