% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{by_position}
\alias{by_position}
\title{Convert a list of tables separated by sample to a list of tables separated by
genome position}
\usage{
by_position(x)
}
\arguments{
\item{x}{List of matrices or vectors}
}
\value{
List of matrices
}
\description{
Convert a list of tables separated by sample to a list of tables separated by
genome position
}
