% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{calc_mutation_rate_from_tables}
\alias{calc_mutation_rate_from_tables}
\title{Calculate the mutation rate from variant and line info tables}
\usage{
calc_mutation_rate_from_tables(variant_table, line_info)
}
\description{
Calculate the mutation rate from variant and line info tables
}
