% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{call_variant}
\alias{call_variant}
\title{Assign a p-value to a variant}
\usage{
call_variant(mut_cov_matrix)
}
\arguments{
\item{mut_cov_matrix}{Numeric matrix; Rows are genome samples, column 1 is
mutant allele count. Column 2 is sequencing coverage.}
}
\value{
A p-value
}
\description{
Assign a p-value to a variant
}
