% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_test.R
\name{call_variants}
\alias{call_variants}
\title{Should be a script}
\usage{
call_variants(og_bams, rot_bams, seq_err_rates)
}
\arguments{
\item{og_bams}{List of bam files from alignment to the unrotated reference
sequences}

\item{rot_bams}{List of bam files from alignment to the rotated reference
sequences}

\item{seq_err_rates}{List of sequencing error rate estimates for each sample}
}
\description{
Should be a script
}
