% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{compare_all_mutation_rates}
\alias{compare_all_mutation_rates}
\title{Bootstrap mutation rate difference for all possible group comparisons in 'by'}
\usage{
compare_all_mutation_rates(variant_table, line_info, by, within = NULL,
  reps = 1e+05)
}
\arguments{
\item{variant_table}{A variant table}

\item{line_info}{A line info table}

\item{by}{Column name in variant_table in which groups can be found}

\item{within}{Column name in variant_table; If specified the comparisons in
'by' are made separately for each group in 'within'.}

\item{group1, group2}{Character; Group IDs in variant_table}
}
\description{
Bootstrap mutation rate difference for all possible group comparisons in 'by'
}
