% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{compare_mutation_rates_within}
\alias{compare_mutation_rates_within}
\title{Helper function called by 'compare_all_mutation_rates'. This function in turn
just repeatedly calls the 2-group bootstrap test function for each pairwise
combination of levels in 'by'}
\usage{
compare_mutation_rates_within(variant_table, line_info, by, reps)
}
\description{
Helper function called by 'compare_all_mutation_rates'. This function in turn
just repeatedly calls the 2-group bootstrap test function for each pairwise
combination of levels in 'by'
}
