% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{compute_all_strand_bias}
\alias{compute_all_strand_bias}
\title{Calculate strand bias for each genome position using fisher tests}
\usage{
compute_all_strand_bias(stranded_piles, consensus, mutant_consensus,
  mutant_samples)
}
\arguments{
\item{stranded_piles}{data.table; Allele count pileups separated by strand}

\item{consensus}{character; The consensus sequence for the samples}

\item{mutant_consensus}{character; An alternate allele consensus sequence
produced by create_mutant_consensus}

\item{mutant_samples}{numeric vector; which sample is a potential mutant for
each genome position}
}
\value{
Numeric vector of p-values
}
\description{
Calculate strand bias for each genome position using fisher tests
}
