% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{compute_allele_frequency}
\alias{compute_allele_frequency}
\title{Calculate an allele's frequency from a vector of allele counts}
\usage{
compute_allele_frequency(counts, allele)
}
\arguments{
\item{counts}{Numeric vector; Allele counts for a single genome position and
sample}

\item{allele}{Character; The target allele; One of {A, C, G, T, +, -}}
}
\value{
Numeric
}
\description{
Calculate an allele's frequency from a vector of allele counts
}
