% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{compute_mutant_allele_frequencies}
\alias{compute_mutant_allele_frequencies}
\title{Calculate mutant allele frequency at each genome position for a sample}
\usage{
compute_mutant_allele_frequencies(pile, mutant_consensus)
}
\arguments{
\item{pile}{A data.table; allele counts at each genome position}

\item{mutant_consensus;}{An alternate allele consensus sequence produced by
create_mutant_consensus}
}
\value{
Numeric vector
}
\description{
Calculate mutant allele frequency at each genome position for a sample
}
