% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{compute_split_indices}
\alias{compute_split_indices}
\title{Compute the indices for splicing the 'original' and 'rotated' pileups}
\usage{
compute_split_indices(sequence_length)
}
\arguments{
\item{sequence_length}{Integer; The length of the reference sequence}
}
\value{
List, length=3; First and third list items are indices from 'rotated'
pileup. Second item is indices into 'original' pileup.
}
\description{
Compute the indices for splicing the 'original' and 'rotated' pileups
}
