% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{compute_strand_bias}
\alias{compute_strand_bias}
\title{Calculate strand bias for a single genome position using fisher exact test}
\usage{
compute_strand_bias(counts, wild_type_allele, mutant_allele)
}
\arguments{
\item{counts}{Numeric vector; Allele counts for a single genome position and
sample}

\item{wild_type_allele}{Character; The wild type allele for this position}

\item{mutant_allele}{Character; The wild type allele for this position}
}
\value{
Numeric; A p-value
}
\description{
Calculate strand bias for a single genome position using fisher exact test
}
