% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{construct_pileup}
\alias{construct_pileup}
\title{Create an allele count pileup with RSamtools}
\usage{
construct_pileup(bam, bp, min_base_quality = 30,
  distinguish_strands = FALSE)
}
\arguments{
\item{bam}{Path to a .bam file}

\item{bp}{Number of base pairs in the mitochondrial reference sequence.}

\item{min_base_quality}{All read positions with phred score <
min_base_quality will be excluded}

\item{distinguish_strands;}{If TRUE, base counts on the +/- strands are
counted separately}
}
\value{
A matrix
}
\description{
Create an allele count pileup with RSamtools
}
