% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{construct_spliced_pileup}
\alias{construct_spliced_pileup}
\title{Splice the original and rotated BAM files into a single allele count table}
\usage{
construct_spliced_pileup(og_bam, rot_bam, bp, min_base_quality = 30,
  distinguish_strands = FALSE)
}
\arguments{
\item{og_bam}{.bam files from alignment to the unrotated reference
sequences}

\item{rot_bam}{.bam files from alignment to the rotated reference
sequences}

\item{bp}{Number of base pairs in the mitochondrial reference sequence.}

\item{min_base_quality}{Numeric; minimum Phred base quality for base to be
included in table}

\item{distinguish_strands;}{If TRUE base counts on the +/- strands are
counted separately}
}
\value{
A data.frame
}
\description{
Splice the original and rotated BAM files into a single allele count table
}
