% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{convert_to_mut_wt_counts}
\alias{convert_to_mut_wt_counts}
\title{Convert an allele count pileup to a table of mutant vs. wild-type counts}
\usage{
convert_to_mut_wt_counts(pile, mutant_consensus)
}
\arguments{
\item{pile}{A matrix; allele counts at each genome position}

\item{mutant_consensus;}{An alternate allele consensus sequence produced by
create_mutant_consensus}
}
\value{
An nx2 matrix, where columns are mutant and wild-type allele counts,
        and rows are genome positions.
}
\description{
Convert an allele count pileup to a table of mutant vs. wild-type counts
}
