% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{convert_to_vcf_format}
\alias{convert_to_vcf_format}
\title{Convert a variant table to VCF format}
\usage{
convert_to_vcf_format(var_table, species)
}
\arguments{
\item{var_table}{Variant table}

\item{species}{The Daphnia species ('magna', 'pulex')}
}
\value{
A data.frame in VCF format
}
\description{
Convert a variant table to VCF format
}
