% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{create_consensus}
\alias{create_consensus}
\title{Create a consensus sequence from a set of nucleotide count pileups}
\usage{
create_consensus(piles, fasta = NULL)
}
\arguments{
\item{piles}{List of data.table; Each list item contains allele counts at
each genome position for a single sample.}

\item{fasta}{File path as character; If given, consensus sequence will be
written to the path}
}
\value{
Character vector; The multisample consensus sequence as a string
}
\description{
Create a consensus sequence from a set of nucleotide count pileups
}
