% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{create_maf_table}
\alias{create_maf_table}
\title{Generate a table of minor allele frequencies}
\usage{
create_maf_table(pile, consensus)
}
\arguments{
\item{pile}{- A data.table; allele counts at each genome position}

\item{consensus}{character; The consensus sequence for the samples}
}
\value{
data.table; Two columns - minor allele and minor allele frequency
}
\description{
Generate a table of minor allele frequencies
}
