% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{create_mut_consensus}
\alias{create_mut_consensus}
\title{Create a sequence of the consensus alternate (mutant) alleles at each genome
position among multiple samples.}
\usage{
create_mut_consensus(piles, consensus)
}
\arguments{
\item{piles}{List of data.table; Each list item contains allele counts at
each genome position for a single sample.}

\item{consensus}{character; The consensus sequence for the samples}
}
\value{
character; The alternate consensus sequence
}
\description{
The alternate allele is the non-consensus base with the highest frequency at
a locus within a single sample. These alternate alleles may be true mutants
or might just be sequencing error.
}
