% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{create_pileup}
\alias{create_pileup}
\title{Create an allele count pileup with RSamtools}
\usage{
create_pileup(bam, min_base_quality = 30, distinguish_strands = FALSE,
  min_nucleotide_depth = 1)
}
\arguments{
\item{bam}{Path to a .bam file}

\item{min_base_quality}{All read positions with phred score <
min_base_quality will be excluded}

\item{distinguish_strands;}{If TRUE, base counts on the +/- strands are
counted separately}

\item{min_nucleotide_depth;}{Minimum allele count to be included in pileup}
}
\description{
Create an allele count pileup with RSamtools
}
