% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dloop_sanger.R
\name{create_sanger_consensus}
\alias{create_sanger_consensus}
\title{Create a single consensus sequence from forward and reverse sequences}
\usage{
create_sanger_consensus(fwd, rev)
}
\arguments{
\item{fwd, rev}{Forward and reverse chromatograms}
}
\description{
Create a single consensus sequence from forward and reverse sequences
}
