% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{create_spliced_count_table}
\alias{create_spliced_count_table}
\title{Splice the original and rotated BAM files into a single allele count table}
\usage{
create_spliced_count_table(og, rot, min_base_quality = 30,
  distinguish_strands = FALSE)
}
\arguments{
\item{og}{Character vector; filepath to the original (og) .bam file}

\item{rot}{Character vector; filepath to the rotated (rot) .bam file}

\item{min_base_quality}{Numeric; minimum Phred base quality for base to be
included in table}

\item{distinguish_strands;}{If TRUE base counts on the +/- strands are
counted separately}
}
\value{
A data.frame
}
\description{
Splice the original and rotated BAM files into a single allele count table
}
