% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{destrand}
\alias{destrand}
\title{Consolidate strand counts in a strand-split allele count table}
\usage{
destrand(counts)
}
\arguments{
\item{counts}{data.table; Table with nucleotide and indel counts for each
genome position split by strand}
}
\value{
data.table; A allele count table with the stranded counts summed
}
\description{
Consolidate strand counts in a strand-split allele count table
}
