% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{determine_unique}
\alias{determine_unique}
\title{Determine which sites are potientially mutant in only one sample}
\usage{
determine_unique(mut_cov_counts, seq_error_rates)
}
\arguments{
\item{mut_cov_counts}{List of matrices; Each matrix has two columns,
first column is mutant allele counts, second column is
total sequencing coverage. Rows are genome positions}

\item{seq_error_rates}{Numeric vector; Sequencing error rate estimates for
each sample}
}
\value{
Logical vector
}
\description{
Determine which sites are potientially mutant in only one sample
}
