% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dloop_sanger.R
\name{.pair_chromatograms}
\alias{.pair_chromatograms}
\title{Pair the chromatogram by sample}
\usage{
.pair_chromatograms(chromatograms)
}
\arguments{
\item{chromatograms}{A list of chromatograms}
}
\value{
A named list of lists. First list is forward sequences,
        second is reverse.
}
\description{
Pair the chromatogram by sample
}
