% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{downsample_pileup}
\alias{downsample_pileup}
\title{Downsample a pileup to an even and uniform sequencing depth.}
\usage{
downsample_pileup(pileup, depth)
}
\description{
Downsample a pileup to an even and uniform sequencing depth.
}
