% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{downsample_vector}
\alias{downsample_vector}
\title{Downsample a vector of allele counts to a target sequencing depth.}
\usage{
downsample_vector(vec, depth)
}
\description{
Downsample a vector of allele counts to a target sequencing depth.
}
