% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{find_indel}
\alias{find_indel}
\title{Find a multinucleotide indel in a table of variants}
\usage{
find_indel(test_table)
}
\arguments{
\item{test_table}{data.frame; A table of variant information, rows are genome
positions}
}
\value{
Numeric vector; row indices of a multinucleotide indel. NULL if no
        indels can be found.
}
\description{
Find a multinucleotide indel in a table of variants
}
