% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{find_runs}
\alias{find_runs}
\title{Find runs of consecutive values with length > 1}
\usage{
find_runs(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
List of numeric vectors
}
\description{
Find runs of consecutive values with length > 1
}
