% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{get_allele_counts}
\alias{get_allele_counts}
\title{Get the allele counts for a set alleles}
\usage{
get_allele_counts(pile, alleles)
}
\arguments{
\item{pile}{A data.table; allele counts at each genome position}

\item{alleles}{Character vector}
}
\value{
Numeric vector
}
\description{
Get the allele counts for a set alleles
}
