% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{get_coverage}
\alias{get_coverage}
\title{Get sequencing coverage for each genome position}
\usage{
get_coverage(pile)
}
\arguments{
\item{pile}{matrix; allele counts at each genome position}
}
\value{
A numeric vector
}
\description{
Get sequencing coverage for each genome position
}
