% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{get_highest_frequency_allele}
\alias{get_highest_frequency_allele}
\title{Get the highest frequency allele in a vector of allele counts
WARNING: Returns the first value if tied}
\usage{
get_highest_frequency_allele(counts)
}
\arguments{
\item{counts}{Numeric vector; Allele counts for a single genome position and
sample}
}
\description{
Get the highest frequency allele in a vector of allele counts
WARNING: Returns the first value if tied
}
