% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{get_minor_allele}
\alias{get_minor_allele}
\title{Get the allele with the second highest frequency for a single sample
WARNING: Returns the first value if tied}
\usage{
get_minor_allele(counts, major_allele)
}
\arguments{
\item{counts}{Numeric vector; Allele counts for a single genome position and
sample}

\item{major_allele}{The major allele at this genome position}
}
\value{
Character; One of {A, C, G, T, +, -}
}
\description{
Get the allele with the second highest frequency for a single sample
WARNING: Returns the first value if tied
}
