% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{get_mut_af}
\alias{get_mut_af}
\title{Get mutant allele frequency at each genome position}
\usage{
get_mut_af(pile, mut_consensus)
}
\arguments{
\item{pile}{- A data.table; allele counts at each genome position}

\item{mut_consensus;}{An alternate allele consensus sequence produced by
`create_mut_consensus`}

\item{Numeric}{vector}
}
\description{
Get mutant allele frequency at each genome position
}
