% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{get_mut_counts}
\alias{get_mut_counts}
\title{Get mutant allele counts at each genome position}
\usage{
get_mut_counts(pile, mut_consensus)
}
\arguments{
\item{pile}{- A data.table; allele counts at each genome position}

\item{mut_consensus;}{An alternate allele consensus sequence produced by
`create_mut_consensus`}
}
\value{
Numeric vector
}
\description{
Get mutant allele counts at each genome position
}
