% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{get_mutation_class}
\alias{get_mutation_class}
\title{Get the type of mutation}
\usage{
get_mutation_class(mutant_allele)
}
\arguments{
\item{allele}{Character; Vector of alleles, {A, C, G, T, +, -}}
}
\value{
Character; Vector of {snv, insertion, deletion}
}
\description{
Get the type of mutation
}
