% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{highest_freq}
\alias{highest_freq}
\title{Get the highest frequency allele in a vector of allele counts}
\usage{
highest_freq(v)
}
\description{
Get the highest frequency allele in a vector of allele counts
}
