% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{indexed_mean_af}
\alias{indexed_mean_af}
\title{Bootstrap statistic; take the mean of the mutant allele frequencies}
\usage{
indexed_mean_af(data, indices)
}
\description{
Bootstrap statistic; take the mean of the mutant allele frequencies
}
