% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{indexed_mutation_rate}
\alias{indexed_mutation_rate}
\title{Bootstrap statistic; calculate the mutation rate}
\usage{
indexed_mutation_rate(dat, indices)
}
\description{
Bootstrap statistic; calculate the mutation rate
}
