% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{is_unique}
\alias{is_unique}
\title{Return TRUE if only one of a list of p-values is < 0.05}
\usage{
is_unique(ps)
}
\description{
Return TRUE if only one of a list of p-values is < 0.05
}
