% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dloop_sanger.R
\name{main}
\alias{main}
\title{Main function}
\usage{
main(fwd, rev, name)
}
\arguments{
\item{fwd, rev}{Forward and reverse chromatograms}

\item{name}{Name of the sample (for writing fasta ID)}
}
\description{
Main function
}
