% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{major_allele}
\alias{major_allele}
\title{Get highest frequency allele for each position from pileup
WARNING: Returns the first value if tied}
\usage{
major_allele(pile)
}
\arguments{
\item{pile}{A pileup}
}
\value{
Character; Vector of major allele names
}
\description{
Get highest frequency allele for each position from pileup
WARNING: Returns the first value if tied
}
