% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{merge_indel}
\alias{merge_indel}
\title{Merge rows of test_table given by indel_idx into a single row.}
\usage{
merge_indel(test_table, indel_idx)
}
\arguments{
\item{test_table}{data.frame; A table of variant information, rows are genome
positions}

\item{indel_idx}{Numeric vector (length > 1)}
}
\value{
A test_table
}
\description{
Merge rows of test_table given by indel_idx into a single row.
}
