% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparisons_correction.R
\name{mult_comparisons_correct}
\alias{mult_comparisons_correct}
\title{Add q-values to a list of test tables}
\usage{
mult_comparisons_correct(test_tables, fdr_level = 0.05)
}
\arguments{
\item{test_tables}{List of test tables}

\item{fdr_level}{False discovery rate to control by}
}
\value{
List of test tables
}
\description{
Add q-values to a list of test tables
}
