% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{mutation_rate}
\alias{mutation_rate}
\title{Calculate the haploid mutation rate}
\usage{
mutation_rate(muts, gen, bp)
}
\arguments{
\item{muts}{numeric; mutant allele frequencies per sample}

\item{gen}{numeric; generation number per sample}

\item{bp}{integer; Number of nucleotides surveyed per sample}
}
\value{
The mutation rate
}
\description{
Calculate the haploid mutation rate
}
