% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rot_base_pos.R
\name{og_to_rot}
\alias{og_to_rot}
\title{Convert indices in original fasta file into rotated indices.
Under rotation, the second half of fasta sequence is moved to start.}
\usage{
og_to_rot(pos, seq_length)
}
\arguments{
\item{pos}{Index in original fasta file}

\item{seq_length}{The length of the fasta sequence}
}
\value{
The rotated index
}
\description{
Convert indices in original fasta file into rotated indices.
Under rotation, the second half of fasta sequence is moved to start.
}
