% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{partition_by_runs}
\alias{partition_by_runs}
\title{Split a vector by its runs of consecutive values
E.g partition_by_runs(c(1, 2, 4, 5)) == list(c(1, 2), c(4, 5))}
\usage{
partition_by_runs(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
List of numeric vectors
}
\description{
Split a vector by its runs of consecutive values
E.g partition_by_runs(c(1, 2, 4, 5)) == list(c(1, 2), c(4, 5))
}
