% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{piles_to_mut_wt}
\alias{piles_to_mut_wt}
\title{Convert allele count pileups to mutant vs wildtype pileups}
\usage{
piles_to_mut_wt(piles, mut_consensus)
}
\arguments{
\item{piles}{List of data.table; Each list item contains allele counts at
each genome position for a single sample.}

\item{mut_consensus;}{An alternate allele consensus sequence produced by
`create_mut_consensus`}
}
\description{
Convert allele count pileups to mutant vs wildtype pileups
}
