% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rot_base_pos.R
\name{rot_to_og}
\alias{rot_to_og}
\title{Convert indices in rotated fasta file into the pre-rotation indices.
Under rotation, the second half of fasta sequence is moved to start.
See rotate_ref.py}
\usage{
rot_to_og(pos, seq_length)
}
\arguments{
\item{pos}{Index in rotated fasta file}

\item{seq_length}{The length of the fasta sequence}
}
\value{
The rotated index
}
\description{
Convert indices in rotated fasta file into the pre-rotation indices.
Under rotation, the second half of fasta sequence is moved to start.
See rotate_ref.py
}
