% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{run_snp_eff}
\alias{run_snp_eff}
\title{R wrapper around snpEff}
\usage{
run_snp_eff(vcf_file, species, jar = "~/src/snpEff/snpEff.jar",
  config = "inst/config/snpEff.config")
}
\arguments{
\item{vcf_file}{VCF file input}

\item{species}{Species corresponding to vcf_file}

\item{jar}{Location of snpEff jar file}

\item{config}{Location of snpEff config file}
}
\value{
Path to the annotated snpEFF VCF file
}
\description{
Assumes snpEff is stored in ~/src. If
}
