% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{splice_pileups}
\alias{splice_pileups}
\title{Splice the 'original' and 'rotated' pileups into a single pileup}
\usage{
splice_pileups(og_pile, rot_pile, rot_ref)
}
\arguments{
\item{og_pile}{Pileup from alignment to the 'original' reference sequence}

\item{rot_pile}{Pileup from alignment to the 'rotated' reference sequence}
}
\value{
A spliced pileup containing the middle halves of both the 'rotated'
and 'original' pileups
}
\description{
Splice the 'original' and 'rotated' pileups into a single pileup
}
