% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{subsample}
\alias{subsample}
\title{Set of functions for creating and manipulating allele count pileups.
Subsample a count vector}
\usage{
subsample(r, cap)
}
\arguments{
\item{r}{An integer vector}

\item{cap}{Integer; Vector will be subsampled such that its sum = cap}
}
\value{
An integer vector
}
\description{
Set of functions for creating and manipulating allele count pileups.
Subsample a count vector
}
