% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparisons_correction.R
\name{subsample_pileup}
\alias{subsample_pileup}
\title{Subsample an allele count pileup}
\usage{
subsample_pileup(pile, cap)
}
\arguments{
\item{pile}{A data.table; allele counts at each genome position}

\item{cap}{Integer; Pileup will be subsampled such that its sum = cap}
}
\value{
A pileup
}
\description{
Subsample an allele count pileup
}
