% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{two_group_boot_mutation_rate_diff}
\alias{two_group_boot_mutation_rate_diff}
\title{Bootstrap difference in means between two groups}
\usage{
two_group_boot_mutation_rate_diff(var_table, line_info, by, group1, group2,
  reps = 1e+05)
}
\arguments{
\item{var_table}{A variant table}

\item{line_info}{A line info table}

\item{by}{Column name in var_table in which groups can be found}

\item{group1, group2}{Character; Group IDs in var_table}

\item{reps}{integer; Number of bootstrap replicates}
}
\value{
A p-value
}
\description{
#' https://stats.stackexchange.com/questions/136661/using-bootstrap-under-h0-to-perform-a-test-for-the-difference-of-two-means-repl
Efron's and Tibshirani's - intro to the bootstrap, page 223
}
