% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{var_table_to_vcf}
\alias{var_table_to_vcf}
\title{Convert a variant table into VCF files split by sample}
\usage{
var_table_to_vcf(var_table)
}
\arguments{
\item{var_table}{A variant table}
}
\description{
VCF files are output in a folder named 'VCF/'. Headers are in 'inst/headers/'
}
