% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_table.R
\name{var_table_to_vcfs}
\alias{var_table_to_vcfs}
\title{Convert a variant table into VCF files split by genotype}
\usage{
var_table_to_vcfs(var_table)
}
\arguments{
\item{var_table}{A variant table}
}
\description{
VCF files are output in a folder named "VCF/". Headers are in "inst/headers/"
}
