% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dloop_sanger.R
\name{write_fasta_sequence}
\alias{write_fasta_sequence}
\title{Write the consensus to fasta file}
\usage{
write_fasta_sequence(sequence, name, filename)
}
\arguments{
\item{sequence}{The fasta sequence}

\item{name}{The name to write to the ID field}

\item{filename}{The path to the output fasta}
}
\description{
Write the consensus to fasta file
}
