% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{add_snp_eff_annotations}
\alias{add_snp_eff_annotations}
\title{Add snpEff annotations to variant table}
\usage{
add_snp_eff_annotations(snp_eff_vcfs, var_table)
}
\arguments{
\item{snp_eff_vcfs}{list of paths to snpEff annotated VCF files}

\item{var_table}{Variant table}
}
\value{
A variant table with snpEff annotations added
}
\description{
Add snpEff annotations to variant table
}
