% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{binom_tests}
\alias{binom_tests}
\title{Carry out a 1-sided binomial test at each genome position for a sample}
\usage{
binom_tests(mut_cov_counts, seq_error_rate)
}
\arguments{
\item{mut_cov_counts}{Data.frame with two columns, first column is mutant
allele counts, second column is total sequencing
coverage. Rows are genome positions}

\item{seq_error_rate}{Numeric; Sequencing error rate estimate}
}
\value{
Vector of p-values
}
\description{
Carry out a 1-sided binomial test at each genome position for a sample
}
