% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{construct_info_field}
\alias{construct_info_field}
\title{Construct the info field of a VCF from a row of the variant table}
\usage{
construct_info_field(var_table_row)
}
\description{
Construct the info field of a VCF from a row of the variant table
}
