% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{convert_to_mut_wt_counts}
\alias{convert_to_mut_wt_counts}
\title{Convert an allele count pileup to a table of mutant vs. wt counts}
\usage{
convert_to_mut_wt_counts(pile, mut_consensus)
}
\arguments{
\item{pile}{- A data.table; allele counts at each genome position}

\item{mut_consensus;}{An alternate allele consensus sequence produced by
`create_mut_consensus`}
}
\value{
An nx2 matrix, where cols are mutant, and wild-type allele counts,
and rows are genome positions.
}
\description{
Convert an allele count pileup to a table of mutant vs. wt counts
}
