% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{create_consensus}
\alias{create_consensus}
\title{Create a consensus sequence from a set of nucleotide count pileups}
\usage{
create_consensus(piles)
}
\arguments{
\item{piles}{List of data.table; Each list item contains allele counts at
each genome position for a single sample.}
}
\value{
Character vector; The multisample consensus sequence as a string
}
\description{
Create a consensus sequence from a set of nucleotide count pileups
}
