% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileups.R
\name{destrand}
\alias{destrand}
\title{Consolidate strand counts in a strand-split allele count table}
\usage{
destrand(pile)
}
\arguments{
\item{pile}{matrix; Table with nucleotide and indel pile for each
genome position split by strand}
}
\value{
Matrix; A allele count table with the stranded pile summed
}
\description{
Consolidate strand counts in a strand-split allele count table
}
