% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{get_afs_by_sample}
\alias{get_afs_by_sample}
\title{Get mutant_allele frequencies for a sample}
\usage{
get_afs_by_sample(sample, variant_table)
}
\description{
Get mutant_allele frequencies for a sample
}
