% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{get_highest_frequency_allele}
\alias{get_highest_frequency_allele}
\title{Get the highest frequency allele in a vector of allele counts
WARNING: Returns the first value if tied}
\usage{
get_highest_frequency_allele(counts)
}
\arguments{
\item{counts}{numeric vector; allele counts for a single genome position and
sample}
}
\value{
Character; One of {A, C, G, T, +, -}
}
\description{
Get the highest frequency allele in a vector of allele counts
WARNING: Returns the first value if tied
}
