% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_variant_calling.R
\name{get_minor_alleles}
\alias{get_minor_alleles}
\title{Get the alleles with the second highest frequency at each position}
\usage{
get_minor_alleles(pile, consensus)
}
\arguments{
\item{pile}{matrix; allele counts at each genome position}

\item{consensus}{The consensus sequence for the samples}
}
\value{
Character; Vector of minor alleles, each allele one of
        {A, C, G, T, +, -}
}
\description{
Get the alleles with the second highest frequency at each position
}
