% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_tools.R
\name{major_allele}
\alias{major_allele}
\title{Get the major allele at each genome position for a single sample}
\usage{
major_allele(pile)
}
\arguments{
\item{pile}{- A data.table; allele counts at each genome position}
}
\value{
Character; Vector of major alleles
}
\description{
WARNING: Returns the first value if tied
}
