% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_rate.R
\name{mutation_rate}
\alias{mutation_rate}
\title{Calculate the haploid mutation rate}
\usage{
mutation_rate(muts, gen, nuc)
}
\arguments{
\item{muts}{numeric; mutant allele frequencies}

\item{gen}{numeric; Mean generation number}

\item{nuc}{integer; Number of nucleotides surveyed}
}
\value{
The mutation rate
}
\description{
Calculate the haploid mutation rate
}
