% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sample_mutation_calling.R
\name{piles_to_mut_wt}
\alias{piles_to_mut_wt}
\title{Convert allele count pileups to mutant vs wildtype pileups}
\usage{
piles_to_mut_wt(piles, mutant_consensus)
}
\arguments{
\item{piles}{list of data.table; each list item contains allele counts at
each genome position for a single sample.}

\item{mutant_consensus;}{An alternate allele consensus sequence produced by
`create_mutant_consensus`}
}
\description{
Convert allele count pileups to mutant vs wildtype pileups
}
