% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_annotation.R
\name{run_snp_eff}
\alias{run_snp_eff}
\title{Wrapper around snpEff}
\usage{
run_snp_eff(vcf_filename, species, snpeff_config)
}
\arguments{
\item{vcf_filename}{VCF file input}

\item{species}{Species corresponding to vcf_file}

\item{config}{Location of snpEff config file}
}
\value{
Path to the annotated snpEFF VCF file
}
\description{
Wrapper around snpEff
}
