% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_comparisons_correction.R
\name{subsample}
\alias{subsample}
\title{Subsample a count vector}
\usage{
subsample(r, cap)
}
\arguments{
\item{r}{An integer vector}

\item{cap}{Integer; Vector will be subsampled such that its sum = cap}
}
\value{
An integer vector
}
\description{
Subsample a count vector
}
